requirements.txt
fastapi>=0.70.0
tensorflow-cpu>=2.0.0
uvicorn>=0.15.0
python-multipart